cv-frontend-updated/requirements.txt
streamlit>=1.35
pillow>=10
numpy>=1
plotly>=5
opencv-python>=4.8
pandas>=2.0
hypothesis>=6
pytest>=8
pytest-mock>=3
ultralytics>=8.0
scikit-learn>=1.3
